+++ CHANGELOG.rst
@@ -16,6 +16,8 @@ Version 4.2
 * Fixed issue where fastqc was not logging stderr to analysis_quaility.log
 * Renamed run_standard_stable4.pl -> run_standard.pl
 * Removed duplicated run_standard_*.pl scripts
+* Pipeline no longer creates absolute path'd symlinks which makes the output
+  directory portable
 
 Version 4.1
 -----------